% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCountsDF.R
\name{buildCountsDF}
\alias{buildCountsDF}
\title{Compute the DESeq2 input counts table}
\usage{
buildCountsDF(df_with_all_events, verbose = 0)
}
\arguments{
\item{df_with_all_events}{\code{List} containing data frames (one replicate in one
data frame) with the maturation state of each read at each maturation site
\itemize{
\item ideally output by \code{\link[=getStates]{getStates()}}.
}}

\item{verbose}{\code{Integer} indicating the level of wordiness:
\itemize{
\item 0 = muted execution
\item 1 = some information about the execution is displayed in the console
}}
}
\value{
A \code{Data.frame} with the number of reads in each replicate having a
given maturation state for all pair of sites.
}
\description{
\code{buildCountsDF()} takes as input a data frame with informations about reads
maturation state at each site (as output by
\code{\link[=getStates]{getStates()}}) and returns a data
frame ready to be fed to DESeq2 to obtain comaturations.

One row in the output data frame corresponds to a pair of maturation events.

One column in the output data frame corresponds to a combination of the
maturation state at each site and a replicate.
}
\examples{


buildCountsDF_output <- buildCountsDF(df_with_all_events)
counts               <- buildCountsDF_output[[1]]
conditions           <- buildCountsDF_output[[2]]

}
