% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExrem.R
\name{getExtrem}
\alias{getExtrem}
\title{Store strand, extremities and length of a set of reads in a data frame}
\usage{
getExtrem(all_reads)
}
\arguments{
\item{all_reads}{\code{List} containing \code{GRanges} objects (one replicate in one
object) - ideally output by \code{\link[=loadReads]{loadReads()}}.}
}
\value{
A \code{List} of data frames with reads informations.
}
\description{
\code{getExtrem()} takes as input a set of reads (as output by \code{\link[=loadReads]{loadReads()}}) and
returns a list of data frames giving informations about the reads:
\itemize{
\item strand
\item start and end positions
\item length
}
}
\examples{

path_to_all_reads <- system.file("extdata",
                                 "loadReads_output_ex.rds",
                                 package = "comaturationTracker")

all_reads         <- readRDS(path_to_all_reads)

all_extremities   <- getExtrem(all_reads)

}
