% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatur.R
\name{getMatur}
\alias{getMatur}
\title{Retrieve the maturation states of a set of reads for a set of maturations}
\usage{
getMatur(all_reads, path_to_maturation_sites)
}
\arguments{
\item{all_reads}{\code{List} containing \code{GRanges} objects (one replicate in one
object) - ideally output by \code{\link[=loadReads]{loadReads()}}.}

\item{path_to_maturation_sites}{\code{String} containing the path to the folder
with all maturation sites files, \emph{e.g.} in gff file format. The type must
be \code{editing} or \code{intron}.}
}
\value{
A \code{List} of data frames with booleans indicating whether or not
reads are matured for each site (one data frame per replicate).
}
\description{
\code{getMatur()} takes as input a set of reads (as output by \code{\link[=loadReads]{loadReads()}}) and
a set of maturation events (e.g. in \code{gff} format) and returns a data frame
giving the maturation state of each read at all the maturation events sites.
}
\examples{

path_to_all_reads        <- system.file("extdata",
                                        "loadReads_output_ex.rds",
                                        package = "comaturationTracker")

all_reads                <- readRDS(path_to_all_reads)

path_to_maturation_sites <- system.file("extdata",
                                        "sites_files",
                                        package = "comaturationTracker")

maturation_states        <- getMatur(all_reads,
                                     path_to_maturation_sites)

}
