% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReads.R
\name{loadReads}
\alias{loadReads}
\title{Retrieve and format reads}
\usage{
loadReads(path, max_length)
}
\arguments{
\item{path}{\code{String} containing the path to the BAM files.}

\item{max_length}{\code{Integer} indicating the maximal authorized length for a
read. Reads longer than this value will be removed. Default to \code{Inf}.}
}
\value{
A (list of) \code{GRanges} object with all the reads.
}
\description{
\code{loadReads()} applies some formatting on BAM files before further dependency
analysis - it offers to remove:
\itemize{
\item reads longer than a provided length
\item reads not mapping any maturation sites
\item reads mapping outside of a given range
\item reads mapping multiple locations
}
}
\examples{

loadReads(1, 1)
loadReads(10, 1)

}
